refactor to move ssh commands to cmds.ts
Move all the SSH command calling into cmds.ts file.
