+++ src/ssh.ts
@@ -1,12 +1,10 @@
 import * as crypto from 'crypto';
 import * as fs from 'fs';
 import * as path from 'path';
-import * as os from 'os';
 import * as core from '@actions/core';
-import { exec, getExecOutput } from '@actions/exec';
 import * as io from '@actions/io';
 import SSHConfig from 'ssh-config';
-import { IGitCmd } from './cmds';
+import { IGitCmd, ISshCmd } from './cmds';
 
 interface MappedHostSaveState {
   mapped_host: string;
@@ -103,114 +101,6 @@ export async function cleanupDeployKeys(gitCmd: IGitCmd): Promise<void> {
   }
 }
 
-export interface ISshCmd {
-  getDotSshPath(): Promise<string>;
-  listKeys(): Promise<PubKey[]>;
-  loadPrivateKeys(keys: string[]): Promise<void>;
-  startAgent(): Promise<void>;
-  killAgent(): Promise<void>;
-}
-
-export async function createSshCmd(): Promise<ISshCmd> {
-  return await SshCmd.createSshCmd();
-}
-
-class SshCmd {
-  private sshAddPath = '';
-  private sshAgentPath = '';
-  private dotSshPath = '';
-
-  // Private constructor; use createSshCmd()
-  private constructor() {}
-
-  static async createSshCmd(): Promise<SshCmd> {
-    const ret = new SshCmd();
-    ret.sshAddPath = await io.which('ssh-add', true);
-    ret.sshAgentPath = await io.which('ssh-agent', true);
-    return ret;
-  }
-
-  async listKeys(): Promise<PubKey[]> {
-    // list current public key identities
-    core.info(`Running ${this.sshAddPath} -L`);
-    const { exitCode, stdout } = await getExecOutput(
-      `"${this.sshAddPath}"`,
-      ['-L'],
-      {
-        ignoreReturnCode: true,
-        silent: true,
-      },
-    );
-
-    if (exitCode > 1 || exitCode < 0) {
-      throw new Error(`Failed to run ${this.sshAddPath} -L`);
-    }
-
-    // take the output and split it on each new line
-    const lines = stdout.trim().split(/\r?\n/);
-    // we'll build up a list of key data to return
-    const ret: PubKey[] = [];
-    for (const identity of lines) {
-      // the parts of the identity are split by a space
-      const parts = identity.trim().split(' ');
-      const pubKey = { algo: parts[0], key: parts[1], comment: parts[2] };
-      ret.push(pubKey);
-    }
-
-    return ret;
-  }
-
-  async loadPrivateKeys(keys: string[]): Promise<void> {
-    core.debug(`Running ssh-add for each key`);
-    for (const key of keys) {
-      await exec(`"${this.sshAddPath}"`, ['-'], {
-        input: Buffer.from(`${key}\n`),
-        ignoreReturnCode: true,
-      });
-    }
-  }
-
-  async startAgent(): Promise<void> {
-    // start up ssh-agent
-    core.info(`Running ${this.sshAgentPath}`);
-    const { stdout } = await getExecOutput(`"${this.sshAgentPath}"`, []);
-
-    // grab up the output as lines
-    const lines = stdout.trim().split(/\r?\n/);
-    for (const line of lines) {
-      const parts = line.match(
-        /^(SSH_AUTH_SOCK|SSH_AGENT_PID)=(.*); export \1/,
-      );
-
-      if (!parts) {
-        continue;
-      }
-
-      // export this data for future steps
-      core.exportVariable(parts[1], parts[2]);
-    }
-  }
-
-  async killAgent(): Promise<void> {
-    await getExecOutput(`"${this.sshAgentPath}"`, ['-k']);
-  }
-
-  async getDotSshPath(): Promise<string> {
-    // realistically we want this to be a temporary file
-    // that we don't tinker with the system setup
-    // it will also allow this to be exposed into a Docker
-    // build context to function correctly
-    if (this.dotSshPath === '') {
-      const dotSshPath = path.join(os.homedir(), '.ssh');
-      await io.mkdirP(dotSshPath);
-      this.dotSshPath = dotSshPath;
-      return dotSshPath;
-    } else {
-      return this.dotSshPath;
-    }
-  }
-}
-
 interface PubKey {
   algo: string;
   key: string;
@@ -234,6 +124,15 @@ function origRepoUri(key: DeployKey): string {
 
 function mappedRepoUri(key: DeployKey): string {
   return `git@${key.mapped_host}:${key.repo_path}`;
+}
+
+export async function getPublicKeys(ssh: ISshCmd): Promise<PubKey[]> {
+  const lines = await ssh.listKeys();
+  return lines.map(line => {
+    // the parts of the identity are split by a space
+    const parts = line.trim().split(' ');
+    return { algo: parts[0], key: parts[1], comment: parts[2] };
+  });
 }
 
 const OWNER_REPO_MATCH = /\b([\w.]+)[:/]([_.a-z0-9-]+\/[_.a-z0-9-]+)?$/i;
